@@ -6,15 +6,17 @@
 
     <groupId>com.wendaoren</groupId>
     <artifactId>wendaoren-dependencies</artifactId>
-    <version>2.7.1-SNAPSHOT</version>
+    <version>3.2.1-SNAPSHOT</version>
     <packaging>pom</packaging>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <skipTests>true</skipTests>
-        <spring-boot-starter-parent.version>2.7.18</spring-boot-starter-parent.version>
+        <spring-boot-starter-parent.version>3.2.5</spring-boot-starter-parent.version>
+        <jakarta.xml.bind-api.version>4.0.2</jakarta.xml.bind-api.version>
+        <com.wendaoren.version>3.2.1-SNAPSHOT</com.wendaoren.version>
     </properties>
 
     <modules>
