@@ -2,9 +2,9 @@ package com.wendaoren.web.validation.constroins;
 
 import com.wendaoren.web.validation.validator.XssConstraintValidator;
 
-import javax.validation.Constraint;
-import javax.validation.Payload;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+import jakarta.validation.constraints.NotNull;
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
