@@ -6,7 +6,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.util.Assert;
 
-import javax.validation.*;
+import jakarta.validation.*;
 import java.util.Set;
 
 /**
