@@ -2,8 +2,8 @@ package com.wendaoren.web.validation.validator;
 
 import com.wendaoren.web.validation.constroins.NotXss;
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Pattern;
